pyproject.toml
[project]
name = "drive-cycle-calculator"
version = "0.3.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pyarrow>=14.0",
    "duckdb>=0.10",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "scipy>=1.17.1",
    "typer>=0.12.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "geopandas>=1.1.3",
    "plotnine>=0.15.3",
]

[project.scripts]
dcc = "drive_cycle_calculator.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drive_cycle_calculator"]

[tool.hatch.build.targets.sdist]
exclude = ["students"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9",
    "jupyter>=1.1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::RuntimeWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["students"]

[tool.ruff.lint]
select = ["E", "F", "I"]
